requirements.txt
numpy
pandas
python-dotenv
pymongo
certifi
matplotlib
scikit-learn
xgboost

# -e .